Error Message When Enabling Audit Logging (Task 6)
When activating the Audit Logging, our students are often seeing an error message "Fail to opt in, please refresh". So far, we've yet to see it fail to enable Audit Logging, after giving it some time (as noted in the instructions already), but the error message itself is giving cause for concern. Accounting for it in the instructions may be a good way to prevent those concerns.